change swerve from static

+++ src/main/java/frc/robot/RobotMap.java
@@ -88,10 +88,10 @@ public class RobotMap {
 	public static final double WHEEL_CIRCUMFERENCE = Math.PI * WHEEL_DIAMETER;
 
 	// Limit Switches for autoClimb
-	public static final double INPUT_FRONT_TOP_SW = 0; // find actual port for limit switches
-	public static final double INPUT_BACK_TOP_SW = 0; // find actual port for limit switches
-	public static final double INPUT_FRONT_BOTTOM_SW = 0; // find actual port for limit switches
-	public static final double INPUT_BACK_BOTTOM_SW = 0; // find actual port for limit switches
+	public static final int INPUT_FRONT_TOP_SW = 0; // find actual port for limit switches
+	public static final int INPUT_BACK_TOP_SW = 0; // find actual port for limit switches
+	public static final int INPUT_FRONT_BOTTOM_SW = 0; // find actual port for limit switches
+	public static final int INPUT_BACK_BOTTOM_SW = 0; // find actual port for limit switches
 
 	//dimensions of the robot in CM
 	public static final int ROBOT_X_IN_CM = 51;
